+++ python/npcomp/compiler/numpy/importer.py
@@ -10,7 +10,7 @@ import traceback
 
 from _npcomp.mlir import ir
 
-from . import logging
+from ..utils import logging
 from .interfaces import *
 
 __all__ = [
