+++ python/npcomp/compiler/numpy/name_resolver_base.py
