+++ python/npcomp/compiler/numpy/value_coder_base.py
