+++ test/Python/Backend/RefJIT/simple_invoke_numpy.py
@@ -2,11 +2,11 @@
 
 import numpy as np
 
-from npcomp.compiler.backend import refjit
-from npcomp.compiler.frontend import *
-from npcomp.compiler import logging
-from npcomp.compiler import test_config
-from npcomp.compiler.target import *
+from npcomp.compiler.numpy.backend import refjit
+from npcomp.compiler.numpy.frontend import *
+from npcomp.compiler.numpy import test_config
+from npcomp.compiler.numpy.target import *
+from npcomp.compiler.utils import logging
 
 # TODO: This should all exist in a high level API somewhere.
 from _npcomp import mlir
