+++ test/Python/Compiler/Numpy/cpa_type_inference.py
@@ -1,6 +1,6 @@
 # RUN: %PYTHON %s | npcomp-opt -split-input-file -npcomp-cpa-type-inference | FileCheck %s --dump-input=fail
 
-from npcomp.compiler import test_config
+from npcomp.compiler.numpy import test_config
 
 import_global = test_config.create_import_dump_decorator()
