+++ test/Python/Compiler/Numpy/ndarray_inference.py
@@ -1,8 +1,8 @@
 # RUN: %PYTHON %s | npcomp-opt -split-input-file -npcomp-cpa-type-inference -canonicalize | FileCheck %s --dump-input=fail
 
 import numpy as np
-from npcomp.compiler import test_config
-from npcomp.compiler.frontend import EmittedError
+from npcomp.compiler.numpy import test_config
+from npcomp.compiler.numpy.frontend import EmittedError
 
 import_global = test_config.create_import_dump_decorator()
