+++ test/Python/Compiler/Numpy/partial_eval_getattr.py
@@ -2,7 +2,7 @@
 
 import collections
 import math
-from npcomp.compiler import test_config
+from npcomp.compiler.numpy import test_config
 
 import_global = test_config.create_import_dump_decorator()
