+++ test/Python/Compiler/Numpy/resolve_const.py
@@ -1,7 +1,7 @@
 # RUN: %PYTHON %s | npcomp-opt -split-input-file | FileCheck %s --dump-input=fail
 """Module docstring."""
 
-from npcomp.compiler import test_config
+from npcomp.compiler.numpy import test_config
 
 import_global = test_config.create_import_dump_decorator()
